feat: Remove unused `lex_derive` and `token` modules
The changes remove the unused `lex_derive` and `token` modules from the codebase. These modules were not being used and were taking up unnecessary space, so they have been removed to clean up the codebase.
