--- crates/lex_derive/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "lex_derive"
-version = "0.1.0"
-edition = "2021"
-authors = ["Louis Meyer (H1ghBre4k3r) <h1ghbe4k3r@dev.bre4k3r.de>"]
-license = "GPL-3.0"
-readme = "README.md"
-repository = "https://github.com/H1ghBre4k3r/y-parser"
-description = "Some derive macros for creating a lexer"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1.0.40"
-syn = "2.0.106"
