--- crates/lex_derive/README.md
@@ -1,3 +1,0 @@
-# Y Parser Derive
-
-> TBD
