+++ crates/why_lib/Cargo.toml
@@ -11,8 +11,6 @@ repository = "https://github.com/H1ghBre4k3r/y-parser"
 anyhow = "1.0.99"
 colored = "3.0.0"
 inkwell = { version = "0.6.0", features = ["llvm18-1"] }
-lex_derive = { path = "../lex_derive/" }
-regex = "1.11.2"
 serde = { version = "1.0.219", features = ["derive"] }
 sha2 = "0.10.9"
 unescape = "0.1.0"
